// src/app/@api/model/faqDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface FaqDto { 
    answer?: string;
    faqId?: number;
    question?: string;
    status?: FaqDto.StatusEnum;
}
export namespace FaqDto {
    export type StatusEnum = 'PUBLISHED' | 'UNPUBLISHED';
    export const StatusEnum = {
        PUBLISHED: 'PUBLISHED' as StatusEnum,
        UNPUBLISHED: 'UNPUBLISHED' as StatusEnum
    };
}